+++ R/method_homer.R
@@ -8,15 +8,15 @@
 #'
 #' @family Transformation methods
 #' @param mdata A mldr dataset used to train the binary models.
-#' @param base.method A string with the name of the base method. (Default:
-#'  \code{options("utiml.base.method", "SVM")})
+#' @param base.algorithm A string with the name of the base algorithm. (Default:
+#'  \code{options("utiml.base.algorithm", "SVM")})
 #' @param clusters Number maximum of nodes in each level. (Default: 3)
 #' @param method The strategy used to organize the labels (create the
 #'  meta-labels). The options are: "balanced", "clustering" and "random".
 #'    (Default: "balanced").
 #' @param iteration The number max of iterations, used by balanced or clustering
 #'  methods.
-#' @param ... Others arguments passed to the base method for all subproblems
+#' @param ... Others arguments passed to the base algorithm for all subproblems.
 #' @param cores The number of cores to parallelize the training. Values higher
 #'  than 1 require the \pkg{parallel} package. (Default:
 #'  \code{options("utiml.cores", 1)})
@@ -44,7 +44,8 @@
 #' ##Change default configurations
 #' model <- homer(toyml, "RF", clusters=5, method="clustering", iteration=10)
 #' }
-homer <- function (mdata, base.method = getOption("utiml.base.method", "SVM"),
+homer <- function (mdata,
+                   base.algorithm = getOption("utiml.base.algorithm", "SVM"),
                    clusters = 3, method = c("balanced", "clustering", "random"),
                    iteration = 100, ..., cores = getOption("utiml.cores", 1),
                    seed = getOption("utiml.seed", NA)) {
@@ -72,7 +73,7 @@ homer <- function (mdata, base.method = getOption("utiml.base.method", "SVM"),
     set.seed(seed)
   }
 
-  hmodel$models <- buildLabelHierarchy(mdata, base.method, method, clusters,
+  hmodel$models <- buildLabelHierarchy(mdata, base.algorithm, method, clusters,
                                        iteration, ..., cores=cores, seed=seed)
 
   utiml_restore_seed()
@@ -90,7 +91,7 @@ homer <- function (mdata, base.method = getOption("utiml.base.method", "SVM"),
 #'  matrix, data.frame or a mldr object.
 #' @param probability Logical indicating whether class probabilities should be
 #'  returned. (Default: \code{getOption("utiml.use.probs", TRUE)})
-#' @param ... Others arguments passed to the base method prediction for all
+#' @param ... Others arguments passed to the base algorithm prediction for all
 #'   subproblems.
 #' @param cores The number of cores to parallelize the prediction. Values higher
 #'  than 1 require the \pkg{parallel} package. (Default:
@@ -136,12 +137,15 @@ predictLabelHierarchy <- function(node, newdata, ..., cores, seed) {
   bipartition <- as.bipartition(prediction)
   probability <- as.probability(prediction)
 
+  metalabel <- paste(unlist(lapply(node$metalabels, paste, collapse="*")),
+                     collapse="|")
+
   for(i in seq(node$metalabels)) {
     labels <- node$metalabels[[i]]
     if (length(labels) > 1) {
       child <- node$children[[i]]
 
-      indexes <- bipartition[, i] == 1
+      indexes <- bipartition[, i, drop=FALSE] == 1
       if (any(indexes)) {
         prediction <- predictLabelHierarchy(child, newdata[indexes, ], ...,
                                             cores=cores, seed=seed)
@@ -151,7 +155,6 @@ predictLabelHierarchy <- function(node, newdata, ..., cores, seed) {
           dimnames = list(rownames(bipartition), colnames(prediction)))
         )
 
-        new.bip[!indexes, ] <- 0
         new.bip[indexes, colnames(prediction)] <- as.bipartition(prediction)
         bipartition <- cbind(bipartition, new.bip)
 
@@ -165,16 +168,31 @@ predictLabelHierarchy <- function(node, newdata, ..., cores, seed) {
         }
         new.prob[indexes, colnames(prediction)] <- as.probability(prediction)
         probability <- cbind(probability, new.prob)
+      } else {
+        #Predict all instances of the meta-label as negative
+        aux <- do.call(cbind, lapply(labels, function(lbl)
+          bipartition[, i, drop=FALSE]))
+        colnames(aux) <- labels
+        bipartition <- cbind(bipartition, aux)
+
+        aux <- do.call(cbind, lapply(labels, function(lbl)
+          probability[, i, drop=FALSE]))
+        colnames(aux) <- labels
+        probability <- cbind(probability, aux)
       }
+    } else {
+      #Rename the meta-label because it is the label
+      colnames(bipartition)[i] <- colnames(probability)[i] <- labels
     }
   }
 
+ #cat(metalabel, "\n")
   multilabel_prediction(
     bipartition[, node$labels, drop=F], probability[, node$labels, drop=F]
   )
 }
 
-buildLabelHierarchy <- function (mdata, base.method, method, k, it,
+buildLabelHierarchy <- function (mdata, base.algorithm, method, k, it,
                                  ..., cores, seed) {
   node <- list(labels = rownames(mdata$labels), metalabels = list())
 
@@ -183,8 +201,20 @@ buildLabelHierarchy <- function (mdata, base.method, method, k, it,
   newls <- do.call(cbind, lapply(node$metalabels, function (u){
     as.numeric(rowSums(mdata$dataset[, u, drop=FALSE]) > 0)
   }))
-  colnames(newls) <- unlist(lapply(node$metalabels, paste, collapse='@'))
+  colnames(newls) <- paste('meta-lbl-', seq(node$metalabels), sep='')
   rows <- which(rowSums(newls) > 0)
+
+  #Fix meta-label without positive instances
+  if (any(colSums(newls) == 0)) {
+    empty.labels <- colSums(newls) == 0
+    node$metalabels <- c(node$metalabels[!empty.labels],
+                         unlist(node$metalabels[empty.labels]))
+    newls <- do.call(cbind, lapply(node$metalabels, function (u){
+      as.numeric(rowSums(mdata$dataset[, u, drop=FALSE]) > 0)
+    }))
+    colnames(newls) <- paste('meta-lbl-', seq(node$metalabels), sep='')
+    rows <- which(rowSums(newls) > 0)
+  }
 
   ndata <- remove_unique_attributes(mldr_from_dataframe(
     cbind(mdata$dataset[rows, mdata$attributesIndexes], newls[rows,, drop=F]),
@@ -192,15 +222,17 @@ buildLabelHierarchy <- function (mdata, base.method, method, k, it,
     name = mdata$name
   ))
 
+  mtlbl <- paste(sapply(node$metalabels, paste, collapse='*'), collapse="|")
+
   node$attributes <- colnames(ndata$dataset[, ndata$attributesIndexes])
-  node$model <- br(ndata, base.method, ..., cores=cores, seed=seed)
+  node$model <- br(ndata, base.algorithm, ..., cores=cores, seed=seed)
   rm(ndata)
 
   node$children <- lapply(node$metalabels, function (metalabels) {
     if (length(metalabels) > 1) {
       excluded.label <- node$labels[!node$labels %in% metalabels]
       ndata <- remove_unlabeled_instances(remove_labels(mdata, excluded.label))
-      buildLabelHierarchy(ndata, base.method, method, k, it, ...,
+      buildLabelHierarchy(ndata, base.algorithm, method, k, it, ...,
                           cores=cores, seed=seed)
     } else {
       NULL