+++ R/zzz.R
@@ -1,7 +1,7 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.utiml <- list(
-    utiml.base.method = "SVM",
+    utiml.base.algorithm = "SVM",
     utiml.cores = 1,
     utiml.seed = NA,
     utiml.use.probs = TRUE,
@@ -9,6 +9,10 @@
   )
   toset <- !(names(op.utiml) %in% names(op))
   if (any(toset)) options(op.utiml[toset])
+
+  if (!exists('.Random.seed', envir = .GlobalEnv, inherits = FALSE)) {
+    sample(c()) #Force .Random.seed creation
+  }
 
   invisible()
 }